refactor(agent): split tool-use cycle flow and agent registry into modules
ToolUseCycleFlow keeps the flow factory and re-exports while its nodes,
shared schema, and tool-call parsing move to core/flows/toolUseCycle/;
agentRegistry keeps the stateful cache core while entry types, constants,
YAML scanning, remote-agent metadata, and option building move to sibling
modules. Original paths stay the public entry points.
